refactor: move everything in helpers that only concerns the server to a server folder
